test: correct subgraph layout reproduction
